chore: BigInteger 타입을 Long으로 일괄 변경

+++ src/main/java/com/dayaeyak/waiting/domain/controller/WaitingController.java
@@ -14,7 +14,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigInteger;
+import java.lang.Long;
+
 import java.util.Map;
 
 @RestController
@@ -39,7 +40,7 @@ public class WaitingController {
     @GetMapping("/{waitingId}")
     public ResponseEntity<ApiResponse<WaitingResponseDto>> getWaiting(
             @Validated
-            @PathVariable BigInteger waitingId){
+            @PathVariable Long waitingId){
         WaitingResponseDto responseDto = waitingService.getWaiting(waitingId);
         return ApiResponse.success(200, "웨이팅 단건이 조회되었습니다. ", responseDto);
     }
@@ -47,7 +48,7 @@ public class WaitingController {
     // 가게별 웨이팅 목록 조회
     @GetMapping
     public ResponseEntity<ApiResponse<WaitingListResponseDto>> getWaitings(
-            @RequestParam BigInteger restaurantId,
+            @RequestParam Long restaurantId,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "20") int size) {
         WaitingListResponseDto responseDto = waitingService.getWaitings(restaurantId, page, size);
@@ -57,7 +58,7 @@ public class WaitingController {
     // 웨이팅 수정 액션 통합형
     @PostMapping("/{waitingId}/actions/{action}")
     public ResponseEntity<ApiResponse<WaitingUpdateResponseDto>> updateAction(
-            @PathVariable BigInteger waitingId,
+            @PathVariable Long waitingId,
             @PathVariable String action,
             @RequestBody(required = false) Map<String, Object> payload) {
         WaitingUpdateResponseDto responseDto = waitingActionService.updateWaiting(waitingId, action, payload);
@@ -70,7 +71,7 @@ public class WaitingController {
     @DeleteMapping("/{waitingId}")
     public ResponseEntity<ApiResponse<Void>> deleteWaiting(
             @Validated
-            @PathVariable BigInteger waitingId){
+            @PathVariable Long waitingId){
         waitingService.deleteWaiting(waitingId);
         return ApiResponse.success(204, "웨이팅이 삭되었습니다.", null);
     }
@@ -79,7 +80,7 @@ public class WaitingController {
     @DeleteMapping("/all/{restaurantId}")
     public ResponseEntity<ApiResponse<Void>> deleteWaitingAll(
             @Validated
-            @PathVariable BigInteger restaurantId){
+            @PathVariable Long restaurantId){
         waitingService.deleteWaitingAll(restaurantId);
         return ApiResponse.success(204, "웨이팅이 삭되었습니다.", null);
     }
