+++ src/main/java/com/dayaeyak/waiting/domain/dto/request/WaitingCreateRequestDto.java
@@ -2,12 +2,12 @@ package com.dayaeyak.waiting.domain.dto.request;
 
 import lombok.Getter;
 
-import java.math.BigInteger;
+import java.lang.Long;
 
 @Getter
 public class WaitingCreateRequestDto {
-    private BigInteger restaurantId;
-    private BigInteger datesId;
-    private BigInteger userId;
+    private Long restaurantId;
+    private Long datesId;
+    private Long userId;
     private Integer userCount;
 }
