+++ src/main/java/com/dayaeyak/waiting/domain/dto/response/NoShowCreateResponseDto.java
@@ -3,10 +3,10 @@ package com.dayaeyak.waiting.domain.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.math.BigInteger;
+
 
 @Getter
 @AllArgsConstructor
 public class NoShowCreateResponseDto {
-    private BigInteger noShowId;
+    private Long noShowId;
 }
