+++ src/main/java/com/dayaeyak/waiting/domain/dto/response/NoShowResponseDto.java
@@ -3,13 +3,13 @@ package com.dayaeyak.waiting.domain.dto.response;
 import com.dayaeyak.waiting.domain.entity.NoShow;
 import lombok.Builder;
 
-import java.math.BigInteger;
+
 
 @Builder
 public record NoShowResponseDto (
-    BigInteger noShowId,
-    BigInteger restaurantId,
-    BigInteger userId
+    Long noShowId,
+    Long restaurantId,
+    Long userId
 ){
 
     public static NoShowResponseDto from(NoShow noshow){
